# [DirectSound] Fix buffer underrun for Bluetooth audio devices